% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/increment-package.R
\docType{package}
\name{increment-package}
\alias{increment}
\alias{increment-package}
\title{increment: Causal Effects Based on Incremental Propensity Score Interventions}
\description{
Implementation of...
}
\author{
\strong{Maintainer}: Nicholas Williams \email{niw4001@med.cornell.edu} (\href{https://orcid.org/0000-0002-1378-4831}{ORCID}) [copyright holder]

}
\keyword{internal}
