% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z-estimator.R
\name{increment}
\alias{increment}
\title{Incremental Propensity Score Z-Estimator}
\usage{
increment(
  data,
  trt,
  outcome,
  baseline,
  time_vary = NULL,
  delta,
  k = Inf,
  outcome_type = c("binomial", "continuous"),
  id = NULL,
  learners_outcome = NULL,
  learners_trt = NULL,
  folds = 10
)
}
\arguments{
\item{data}{A data frame in wide format containing all necessary variables
for the estimation problem.}

\item{trt}{A vector containing the column names of treatment variables ordered by time.}

\item{outcome}{The column name of the outcome variable.}

\item{baseline}{An optional vector of columns names of baseline covariates to be
included for adjustment at every timepoint.}

\item{time_vary}{A list the same length as the number of time points of observation with
the column names for new time-varying covariates introduced at each time point. The list
should be ordered following the time ordering of the model.}

\item{delta}{A sequence of increment values for the propensity.}

\item{k}{An integer specifying how previous time points should be
used for estimation at the given time point. Default is \code{Inf},
all time points.}

\item{outcome_type}{Outcome variable type (i.e., continuous, binomial).}

\item{id}{An optional column name containing cluster level identifiers.}

\item{learners_outcome}{An optional \code{sl3} learner stack for estimation of the outcome
regression. If not specified, will default to using a generalized linear model.}

\item{learners_trt}{An optional \code{sl3} learner stack for estimation of the exposure
mechanism. If not specified, will default to using a generalized linear model.}

\item{folds}{The number of folds to be used for cross-fitting. Minimum allowable number
is two folds.}
}
\value{
A list of class \code{increment} containing the following components:

\item{estimates}{A data.frame with the same number of rows as \code{delta} containing intervention estimates.}
\item{eif}{A list the same length as delta containing observation values of the influence function.}
\item{mult_cv}{Multiplier bootstrap critical value.}
}
\description{
Incremental Propensity Score Z-Estimator
}
\examples{
# TO DO
}
\references{
Kennedy EH. Nonparametric causal effects based on incremental
propensity score interventions.
\href{https://arxiv.org/abs/1704.00211}{arxiv:1704.00211}
}
